style: resolve linting/prettier issues
