--- src/errors/error.abstract.ts
@@ -1,5 +1,0 @@
-import type { ErrorCode } from './codes'
-
-export abstract class OpenFeatureError extends Error {
-  abstract code: ErrorCode
-}
