+++ src/errors/flag-not-found-error.ts
@@ -1,0 +1,11 @@
+import { OpenFeatureError } from './error-abstract';
+import { ErrorCode } from './codes.js';
+
+export class FlagNotFoundError extends OpenFeatureError {
+  code: ErrorCode;
+  constructor(message?: string) {
+    super(message);
+    Object.setPrototypeOf(this, FlagNotFoundError.prototype);
+    this.code = ErrorCode.FLAG_NOT_FOUND;
+  }
+}
