+++ src/errors/index.ts
@@ -1,4 +1,4 @@
 export * from './codes.js';
-export * from './flag-not-found.error.js';
-export * from './parse.error.js';
-export * from './type-mismatch.error.js';
+export * from './flag-not-found-error.js';
+export * from './parse-error.js';
+export * from './type-mismatch-error.js';
