+++ src/errors/type-mismatch-error.ts
@@ -1,0 +1,11 @@
+import { OpenFeatureError } from './error-abstract';
+import { ErrorCode } from './codes.js';
+
+export class TypeMismatchError extends OpenFeatureError {
+  code: ErrorCode;
+  constructor(message?: string) {
+    super(message);
+    Object.setPrototypeOf(this, TypeMismatchError.prototype);
+    this.code = ErrorCode.TYPE_MISMATCH;
+  }
+}
